requirements-dev.txt
-r requirements.txt
black==22.3.0
flake8==4.0.1
isort==5.10.1
pytest==7.0.0
pytest-asyncio==0.19.0
pytest-cov==3.0.0
pytest-mock==3.7.0
mock==4.0.3
bs4
requests
pyyaml>=5.3.1